Changed Drop Trap from an Entity to an Animated Tile and updated all references.
