+++ src/com/mojang/mojam/entity/building/Turret.java
@@ -6,7 +6,6 @@ import java.util.Set;
 import com.mojang.mojam.MojamComponent;
 import com.mojang.mojam.entity.Bullet;
 import com.mojang.mojam.entity.Entity;
-import com.mojang.mojam.entity.mob.DropTrap;
 import com.mojang.mojam.entity.mob.Mob;
 import com.mojang.mojam.entity.mob.RailDroid;
 import com.mojang.mojam.entity.mob.SpikeTrap;
@@ -77,8 +76,7 @@ public class Turret extends Building implements IEditable {
     		Entity closest = null;
     		double closestDist = 99999999.0f;
     		for (Entity e : entities) {
-    			if (!(e instanceof Mob) || (e instanceof RailDroid && e.team == this.team) || e instanceof Bomb || e instanceof SpikeTrap || 
-    					e instanceof DropTrap)
+    			if (!(e instanceof Mob) || (e instanceof RailDroid && e.team == this.team) || e instanceof Bomb || e instanceof SpikeTrap)
     				continue;
     			if (!((Mob) e).isNotFriendOf(this))
     				continue;
